+++ lib/marv/cli.rb
@@ -14,11 +14,13 @@ module Marv
     end
 
     desc "create DIRECTORY", "Creates a Marv project"
+    long_desc "Creates a new project. Use the layout option to choose a scaffold"
+    option :layout, :type => :string, :default => 'default', :desc => "Name of alternate layout"
     def create(dir)
       theme = {}
       theme[:name] = dir
 
-      project = Marv::Project.create(dir, theme, self)
+      project = Marv::Project.create(dir, theme, self, options[:layout])
     end
 
     desc "link PATH", "Create a symbolic link to the compilation directory"